# apps/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Laravel Commerce API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Same origin as the web application, through its proxy."
        }
    ],
    "paths": {
...
        "/cart/items/{item}": {
            "patch": {
                "operationId": "cart.items.update",
                "tags": [
                    "CartItem"
                ],
                "parameters": [
                    {
                        "name": "item",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SetCartItemQuantityRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`CartResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CartResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "409": {
                        "description": "For sale, but not in the quantity asked for - or no longer for sale at all.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "available": {
                                            "type": [
                                                "integer",
                                                "null"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "message",
                                        "available"
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "cart.items.destroy",
                "tags": [
                    "CartItem"
                ],
                "parameters": [
                    {
                        "name": "item",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`CartResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CartResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/images/{image}": {
            "get": {
                "operationId": "images.show",
                "tags": [
                    "Image"
                ],
                "parameters": [
                    {
                        "name": "image",
                        "in": "path",
                        "required": true,
                        "description": "The image UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "image/webp": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        },
                        "headers": {
                            "Transfer-Encoding": {
                                "required": true,
                                "schema": {
                                    "type": "string",
                                    "enum": [
                                        "chunked"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            }
        },
...
        "/seller/products/{product}": {
            "get": {
                "operationId": "seller.products.show",
                "tags": [
                    "Product"
                ],
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "description": "The product ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`ProductResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ProductResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "patch": {
                "operationId": "seller.products.update",
                "tags": [
                    "Product"
                ],
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "description": "The product ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`ProductResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ProductResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "delete": {
                "operationId": "seller.products.destroy",
                "summary": "A soft delete. The row stays, so the order history that will eventually\nreference it keeps something to point at, and a shopper stops seeing it\nimmediately because every public query excludes trashed rows",
                "tags": [
                    "Product"
                ],
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "description": "The product ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/seller/products/{product}/images/{image}": {
            "patch": {
                "operationId": "seller.products.images.update",
                "description": "Position is a plain integer rather than a move-up/move-down, because a\ngallery is dragged rather than nudged and the client knows where it\ndropped something.",
                "summary": "Reordering, and the text a screen reader will read",
                "tags": [
                    "ProductImage"
                ],
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "description": "The product ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "image",
                        "in": "path",
                        "required": true,
                        "description": "The image UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProductImageRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`ProductImageResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ProductImageResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "delete": {
                "operationId": "seller.products.images.destroy",
                "description": "A hard delete, unlike a product's. Nothing references a photograph the\nway order history references a listing, and an orphaned file on a disk\nnobody can reach is just cost.",
                "summary": "Removes the row **and** the file",
                "tags": [
                    "ProductImage"
                ],
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "description": "The product ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "image",
                        "in": "path",
                        "required": true,
                        "description": "The image UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/seller/products/{product}/variants/{variant}": {
            "patch": {
                "operationId": "seller.products.variants.update",
                "tags": [
                    "ProductVariant"
                ],
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "description": "The product ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "variant",
                        "in": "path",
                        "required": true,
                        "description": "The variant ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateVariantRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`ProductVariantResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ProductVariantResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "delete": {
                "operationId": "seller.products.variants.destroy",
                "summary": "Removing the last variant raises CannotRemoveLastVariantException, which\nrenders as 409: a product with no variants has no price and cannot be\nbought, which is not a state anything downstream is written to handle",
                "tags": [
                    "ProductVariant"
                ],
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "description": "The product ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "variant",
                        "in": "path",
                        "required": true,
                        "description": "The variant ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/shops/{shopSlug}/products/{productSlug}": {
            "get": {
                "operationId": "shops.products.show",
                "tags": [
                    "PublicProduct"
                ],
                "parameters": [
                    {
                        "name": "shopSlug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "productSlug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`PublicProductResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PublicProductResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CartResource": {
                "type": "object",
                "properties": {
                    "item_count": {
                        "type": "integer",
                        "description": "Units, not lines: three of one thing is a cart of three, which is\nwhat the number beside a cart icon means."
                    },
                    "has_unavailable_items": {
                        "type": "boolean",
                        "description": "The answer a checkout button needs, rather than the lines for the\nbrowser to scan."
                    },
                    "shops": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CartShopResource"
                        }
                    }
                },
                "required": [
                    "item_count",
                    "has_unavailable_items",
                    "shops"
                ],
                "title": "CartResource"
            },
...
            "ProductResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "status": {
                        "$ref": "#/components/schemas/ProductStatus"
                    },
                    "currency": {
                        "description": "Read from the shop, because that is the only place it lives\n(ADR 0007). Sent on the product so a client does not have to\nfetch the shop to know what a price means.",
                        "$ref": "#/components/schemas/Currency"
                    },
                    "published_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "variants": {
                        "type": "array",
                        "description": "There is no `price` here, and there will not be. A product with\ntwo sizes has two prices, and inventing a headline figure for it\nwould be the API guessing which one matters.",
                        "items": {
                            "$ref": "#/components/schemas/ProductVariantResource"
                        }
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductImageResource"
                        }
                    },
                    "can_edit": {
                        "type": "boolean"
                    },
                    "can_publish": {
                        "type": "boolean"
                    },
                    "is_public": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "slug",
                    "description",
                    "status",
                    "currency",
                    "published_at",
                    "created_at",
                    "updated_at",
                    "variants",
                    "images",
                    "can_edit",
                    "can_publish",
                    "is_public"
                ],
                "title": "ProductResource"
            },
...
            "SetCartItemQuantityRequest": {
                "type": "object",
                "properties": {
                    "quantity": {
                        "type": "integer",
                        "description": "`min:1`, so zero is a 422 rather than a quiet delete. A quantity\nof nothing is not a quantity, and `DELETE /cart/items/{item}`\nalready removes a line - two ways to do one thing is how they\nend up behaving differently.",
                        "minimum": 1,
                        "maximum": 999
                    }
                },
                "required": [
                    "quantity"
                ],
                "title": "SetCartItemQuantityRequest"
            },
...
            "UpdateProductImageRequest": {
                "type": "object",
                "properties": {
                    "position": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 8
                    },
                    "alt_text": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "Nullable rather than absent-or-string: clearing alternative text\nis a thing somebody may legitimately want to do.",
                        "maxLength": 200
                    }
                },
                "title": "UpdateProductImageRequest"
            },
            "UpdateProductRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 2,
                        "maxLength": 200
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 5000
                    }
                },
                "title": "UpdateProductRequest"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "AuthorizationException": {
                "description": "Authorization error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}